Fix some bugs

+++ database/migrations/2023_12_12_094919_create_food_table.php
@@ -12,9 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create("foods", function (Blueprint $table) {
-            $table->increments('id')->primary;
+            $table->bigIncrements('id')->primary;
             $table->string('name',255);
-            $table->integer('group_id');
+            $table->unsignedBigInteger('group_id');
             $table->string('money',255);
             $table->boolean('is_active')->default(false);
             $table->text('how_to_make');
